Move everything beneath synnax to make importing clearer
When importing the synnax repo as a Guix channel, we need to know the
location of the definitions for the packages. On the command line,
packages are defined by the name provided in the definition, so
previous and future operations SHOULD NOT change, but when using it
from within the code of Guix (to define an OS for instance), the
`(use-modules ...)` import is clearer.