Add desktop.scm, for configuring my big desktop
This needed special customization, because it uses a SuperMicro
SuperServer board, which has on-board IPMI.
In order for the computer to properly shut off, it needs the
`IPMI_POWEROFF` kernel module, which requires several other IPMI
modules.
Only some of these are turned on by default in Guix/Linux-libre. But I
needed more than was originally provided, so I created a new
linux-libre package, which does what I need it to do.