Switch to removable bootloader for desktop
You would think a nearly $1000 motherboard would have a good UEFI. It
probably does, but the way I am using it means the UEFI forgets OS
boot UEFIs if the disk is unplugged, which is weird.

The removable bootloader means the boot boot EFI file is placed in the
EFI-specification location, ensuring the UEFI is always able to find a
boot file.