Add service to override environment variables for SQLite
This particular way of setting an environment variable will set it for
EVERY shell that you can run (hopefully). Instead of making it
specific to Bash or ZSH, the environment variables listed here are
added to the `setup-environment` script in `.guix-home`, which is
expected to be sourced by the *login shell* (See Guix-Home Manual).