stumpwm: add stumpwm without documentation
This is an example for how I can inherit a package and mess around
with JUST the phases in the build process.

Inheritance is as strong as it gets: you start off with a record that
is identical to the inherited one, then only the fields you
specify (below it) are replaced. The replacement is complete, though:
if you write `(inherit foo) (arguments '(blah))`, no trace of foo's
arguments will remain. If you want to mix at the level of phases, use
`substitute-keyword-arguments` and `modify-phases` on the inherited
ones.
You can do this even without inheriting.