scripts: add package building Guile script to fix my desktop monitors
My desktop monitors work strangely on Linux for some reason. Probably
something to do with X11. Sometimes the smaller monitor (the 1600x900
one) refuses to turn on until the resolution is changed. Then, Guix
always places my HDMI monitor (the big one) to the right of the
smaller one, which is physically incorrect.

The fix-desktop-monitors package uses a script which is defined by the
monitor-script symbol. This script is a Guile script which calls out
to the system to make things happen, namely to xrandr. It also prints
what it is doing to STDOUT, so that we can track what it is doing.