sysadmin: cuirass-mailer: use program-file instead
For some reason, this has been working better than computed-file ever
did. I guess using a guile script to do something is not that bad
either.

The reason I went with building the shell script first was because
that is what Guix's Berlin CI servers use. So I wanted to start there
and work my way over to Scheme. But that does not seem to be the best
way forward.