sysadmin: cuirass: system* uses multiple strings, first MUST be prog
(system* "prog" "args"...) requires that the first string be the
command/program to run. This is important to note because when I was
passing all of the arguments to MSMTP as a single string to system*,
system* interpreted that entire string as the command, which
corresponded to a file that obviously did not exist.