deploys: move guix deploy systems to separate directory
The directory field in the .guix-channel file is the list of
directories that guix-pull should evaluate. If this field is not
specified, then all directories and their subdirectories are treated
as Guile modules, attempting to be loaded and evaluated.

This should solve Guix's misunderstanding of the guix-deploy
operating-system syntax error in the list of machines.
(exception syntax-error (value operating-system) (value "invalid field
specifier") (value ((filename . "ci.scm") (line . 86) (column . 7)))
(value (operating-system %system)) (value %system))