deploys: ci: Make dhcp-client machines send their hostname
dhcp-client-service-type does nto allow for changing the configuration
file /etc/dhclient.conf through Guix. Instead, we must use a
simple-service to extend the etc-service-type, writing dhclient.conf
into /etc/. Right now, the file only says for the host-name of the
machine to be sent on all interfaces that perform a DHCP request.