systems: packages: Add nix as a global package
This should make things easier for compatibility. I can leverage Nix
and nixpkgs while I move things to Guix. Some things are much easier
to use from Nix right now, so we just use that.

NOTE: This does NOT make nix available for use on the system! You MUST
use nix-service-type in the system configuration to make that happen.