systems: home: add automatic direnv shell support to bash
direnv changes environment variables for youb ased on the project and
a .envrc file in the base of the project tree. Adding this eval hook
to the bottom of the .bashrc file means Bash will automatically pick
up direnv settings on-the-fly.

https://direnv.net/docs/hook.html

Direnv comes last so that the generated configuration ends up at the
end of the generated .bashrc file. That way direnv is the last thing
to be set up when starting a new shell.