Embed environment variables to expand rather than getenv calls
Remove string-append from define-configuration default values

If you used string-append and getenv, the getenv and string-append
would be run DURING a guix pull. This causes issues where $HOME is
undefined during a pull inside the inferior guix instance, which
causes #f to be returned, and there is a type-error as
(string-append #f "a string") does not work.