services: mail: mu: mu's maildir now accepts a gexp or a string
This is done for a similar reason as the other getenv-default-value
changes in this file. But because the maildir is used in the
activation script, it behaves even more differently than the others,
which are just applied in configuration files, where the environment
variables can be looked up at run-time.

Here, we want the user to be able to provide THEIR home directory as a
default, which means we need to gexp the building of that path so that
the activation script builds the mail path when the activation script
is run, at which point $HOME is defined. If the user prefers though,
they can also just provide the path to the mail directory as a string,
which is passed through like normal.