systems: home: use file-append for fontconfig mcron job
This does nothing but make the fontconfig job's definition faster, as
the path to fc-cache is embedded as a string in the job, rather than
needing to have mcron build the job with string-append.

Realistically, this makes no change to usage, and provides negligible
speedup, but this makes more sense from a task-delegation perspective.
The guix builder SHOULD be able to output the entire path, as all
inputs are known to the builder.