systems: home: Add config file for SBCL to restrict compiler policy
This means that any time I want to compile something and my user is
the one who does it, debug and safety information will be greatly
preserved. See (sbcl) Initialization Files in the SBCL manual for
more. If the code being generated is intended for others, you should
disable the config file with --no-userinit.