systems: website: Add Referrer-Policy header to all responses
When a user navigates to a site via a hyperlink or a website loads an
external resource, browsers inform the destination site of the origin
of the requests through the use of the HTTP Referer (sic) header.

You can change this policy by passing one of the strings corresponding
to one of the directives at this MDN link below.
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Referrer-Policy#directives

https://infosec.mozilla.org/guidelines/web_security#referrer-policy