packages: scripts: Add script to find Nix GC roots on whole disk
This asks the Nix store what are the current GC roots and finds them
on the system. You can use this to find out why items in the store are
not being GC'd.

NOTE: It _does_ filter out "censored" packages which have GC roots.
These are typically because of things done with nix-env or by the root
user, which I don't really want to find with this script anyways.