systems: packages: Remove emacs-org-roam
Emacs 29 introduced the ability to work with sqlite databases through
Emacs Lisp directly while maintaining the speed that we require and
achieved through a specially-compiled SQLite binary. Now that we do
not need the specially-compiled version of sqlite, we no longer need
to depend on Guix to provide the emacs-org-roam package.