systems: avocato: Add a swap file
This is a swapfile that CANNOT be used for hibernation yet! This is
because the swapfile is stored on the encrypted root volume of the
disk. If we wanted to support hibernation, then we would have to set
some kernel command-line parameters, as documented in
(guix)Swap Space.

To make /swapfile, you need to do the following:

$ sudo fallocate -l 64G /swapfile
$ sudo chmod 600 /swapfile
$ sudo mkswap /swapfile

The final mkswap command also tells you the UUID of the file.