services: home: tmux: New home service to generate tmux configuration.
My tmux configuration was getting so long that it was taking over the
home.scm file with just inline strings. So, I factored it out to a
separate service-type. This has the added benefit of pulling the tmux
package into my config ONLY if I opt into the tmux configuration,
rather than always globally installing it. After all, if I go through
the effort to use a home-environment and install tmux there, I want to
use my configuration of it.