systems: home: Add SSH config for EECS machines
This way I no longer need to type in EECS machines' long hostnames
every time I want to connect to them. This will apply to machines like
hanlon, the batman cluster, and the Wilkinson lab.

NOTE: Moore is special and I have a separate SSH config entry for it,
because it is the designated student teaching machine.