packages: resume: Use texlive-updmap.cfg to run hook to update fonts
Without this, all fonts in my resume would be bitmap rather than
vector, which means that zooming in would produce some TERRIBLE
looking fonts.

You need to run texlive-updmap.cfg as a package input to update the
paths to font files so that LaTeX compilers use the Type 1 vector
fonts which do not have this problem.

I included the collections for many fonts and some extras, just to be
on the safe side. If a font ends up mentioned twice (once manually and
once in a collection), I still only end up with one anyways.

Below, I reproduce the a comment from Guix's gnu/packages/tex.scm,
which states WHY texlive-updmap.cfg needs to be called by packages in
the first place.

    Default font map files are updated in a profile hook (see
    `texlive-font-maps' in "profiles.scm"). However, this option is
    not available when building documentation for a package.
    Consequently, this module also provides TEXLIVE-UPDMAP.CFG
    function, which creates a TeX Live tree with font map files
    updates. It should be used exclusively for package definitions, as
    a native input. It is possible to augment that tree, in particular
    with additional font packages.