systems: ci: Update with XCP-ng VM config current running
I have been slowly migrating to XCP-ng for additional stability and
flexibility in my home-lab. This meant that I was tinkering with the
setup quite a bit. To make this easier, I used a separate file that I
was not tracking in Git. Now that I have made the migration to XCP-ng,
I am upgrading that tinker-config to a real thing.

NOTE: (service xe-guest-agent-service-type) is provided by Guix
services, but it has NOT been accepted upstream yet! I am still
tinkering to work out the final kinks (and hopefully get the Rust
version into Guix as well). So you MUST have
xe-guest-agent-service-type defined in Guix's
(gnu services virtualization) before attempting to build this config!