systems: ci: Give CI 2 workers
The default is to give 1 worker, but my resume takes 2000+ seconds to
compile for some reason, which blocks CI for most of the other things
that I care more about (my resume doesn't change as often as my
software...). Plus, everything I am building is not heavy (at least
for now). The 4 cores assigned to the CI VM is more than enough for
what I am doing with 1 worker right now, so I'm hoping that bumping up
to 2 workers gives the CI system a little bit more flexibility when it
comes to building stuff, without causing the house of cards to tumble
down.