deploys: website: Use website DNS name instead of IP address for deploys
This is nothing other than a QoL improvement for me, since (in theory)
the IP address of the VM I use for Guix _could_ change. But that has
never happened, so this is more just to line up with my philosophy
that configurations should be as IP address agnostic as possible.