systems: avocato & lilcato: Add power-profiles-daemon-service-type
Plasma makes use of the power profiles daemon to control the power usage
of the system. An alternative to power-profiles-daemon is tlp.

I am hoping that adding this service-type to my laptops, I can get
better battery life without sacrificing too much performance. The
laptops already throttle their CPUs when not busy, but there may be
other savings that power-profiles-daemon finds that we can take
advantage of.