systems: packages: Add ncurses for tput command for lualatex
LuaLaTeX needs to know output terminal information when you compile a
document with -shell-escape. I speculate this is for color support,
since lualatex can output its errors in color, but I really have no
idea.

I wonder if this was needed for building the initial font cache and
that's it?